Update Fast Exponentiation.cpp

+++ Math/Fast Exponentiation.cpp
@@ -1,4 +1,4 @@
-int exp(int b, int e, int m=MOD){ // O(logE)
+int exp(ll b, int e, int m=MOD){ // O(logE)
     int res=1;
     while(e){
         if(e&1) res=(res*b)%m;
